reubicacion de lista
